Ensure commit has is clean
Previously the commit hash was added regardless if there are uncommitted changes or not. This meant that a commit hash in a build is not strictly bound to the file-state. Now the script also checks if there are uncommitted changes and marks the revision as such. This is still not perfect as the developer can use a poorly configured .gitignore to get changes into the same hash. However, it proves as a guard against carelessness.